requirements.txt
asgiref==3.2.3
astroid==2.3.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.41.0
django-braces==1.14.0
django-crispy-forms==1.9.0
django-google-maps==0.12.1
django-location-field==2.1.0
django-model-utils==4.0.0
django-private-chat==0.2.2
django-registration==3.1
django-registration-redux==2.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
names==0.3.0
oauthlib==3.1.0
Pillow==7.0.0
pycparser==2.19
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
websockets==3.2
wincertstore==0.2
wrapt==1.11.2
